Backend/requirements.txt
# Shelf Bookstore — Python Dependencies
# Install with: pip install -r requirements.txt

# ── Web Framework ──────────────────────────────────────
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1

# ── Database ───────────────────────────────────────────
mysql-connector-python==8.4.0

# ── Authentication ─────────────────────────────────────
bcrypt==4.1.3

# ── NLP / Embeddings ───────────────────────────────────
sentence-transformers==3.0.1
transformers==4.41.2
torch==2.3.0
tokenizers==0.19.1
huggingface-hub==0.23.2

# ── Vector Search ──────────────────────────────────────
faiss-cpu==1.8.0

# ── Numeric ────────────────────────────────────────────
numpy==1.26.4

# ── Utilities ──────────────────────────────────────────
python-dotenv==1.0.1
Werkzeug==3.0.3
click==8.1.7
